Metrics moved to its own app. Each metrics page now built in template rather than riot tag.
