+++ apps/klasses/templates/klasses/wizard/overview.html
@@ -64,7 +64,7 @@
                 <div class="ui grid" style="width: 100%">
                     <div class="row">
                         <div class="content">
-                            <a href="{% url "klasses:klass_metrics" klass_pk=object.pk %}" class="ui basic blue button">
+                            <a href="{% url "metrics:klass" klass_pk=object.pk %}" class="ui basic blue button">
                                 <i class="ui chart bar outline icon"></i>
                                 Metrics
                             </a>
