+++ apps/klasses/urls.py
@@ -8,7 +8,6 @@ urlpatterns = [
     path('create/', views.CreationView.as_view(), name='create_klass'),
     path('edit/<int:klass_pk>', views.EditView.as_view(), name='edit_klass'),
     path('wizard/<int:klass_pk>', views.OverView.as_view(), name='klass_details'),
-    path('wizard/<int:klass_pk>/metrics', views.KlassMetricsView.as_view(), name='klass_metrics'),
     path('wizard/<int:klass_pk>/enroll', views.EnrollmentView.as_view(), name='klass_enrollment'),
     path('wizard/<int:klass_pk>/define_homework', views.DefineHomeworkView.as_view(), name='klass_homework'),
     path('wizard/<int:klass_pk>/grade_homework', views.GradeHomeworkView.as_view(), name='klass_grading'),
